+++ cloudwanderer/storage_connectors/dynamodb.py
@@ -11,7 +11,7 @@ import boto3
 from random import randrange
 from .base_connector import BaseStorageConnector
 from boto3.dynamodb.conditions import Key, Attr, ConditionBase
-from ..cloud_wanderer_resource import CloudWandererResource
+from ..cloud_wanderer_resource import CloudWandererResource, SecondaryAttribute
 from ..aws_urn import AwsUrn
 from ..utils import standardise_data_types
 
@@ -87,7 +87,10 @@ def _dynamodb_items_to_resources(items: Iterable[dict], loader: Callable) -> Ite
     """
     for item_id, group in itertools.groupby(items, lambda x: x['_id']):
         grouped_items = list(group)
-        attributes = [attribute for attribute in grouped_items if attribute['_attr'] != 'BaseResource']
+        attributes = [
+            SecondaryAttribute(name=attribute['_attr'], **attribute)
+            for attribute in grouped_items if attribute['_attr'] != 'BaseResource'
+        ]
         base_resource = next(iter(resource for resource in grouped_items if resource['_attr'] == 'BaseResource'))
         yield CloudWandererResource(
             urn=_urn_from_primary_key(base_resource['_id']),
@@ -170,7 +173,7 @@ class DynamoDbConnector(BaseStorageConnector):
         for secondary_attribute in resource.cloudwanderer_metadata.secondary_attributes:
             self._write_secondary_attribute(
                 urn=resource.urn,
-                attribute_type=secondary_attribute.attribute_name,
+                attribute_type=secondary_attribute.name,
                 secondary_attribute=secondary_attribute
             )