+++ doc_source/conf.py
@@ -29,7 +29,8 @@ release = '0.10.2'
 
 nitpicky = True
 nitpick_ignore = [
-    ('py:exc', 'botocore.exceptions.ClientError')
+    ('py:exc', 'botocore.exceptions.ClientError'),
+    ('py:class', 'botocore.client.ClientCreator')
 ]
 
 
@@ -76,6 +77,9 @@ intersphinx_mapping = {
     'boto3': ('https://boto3.amazonaws.com/v1/documentation/api/latest/', None),
     'botocore': ('https://botocore.amazonaws.com/v1/documentation/api/latest/', None)
 }
+
+# -- Napoleon
+napoleon_include_init_with_doc = True
 
 # -- Autodoc
 add_module_names = False